+++ src/test/java/com/meti/TypePredicateTest.java
@@ -1,5 +1,6 @@
 package com.meti;
 
+import com.meti.predicate.TypePredicate;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
